refactor: split campaign instance event factories
